余弦定理的实现

+++ src/search/model/Dict.java
@@ -4,7 +4,7 @@ public class Dict {
 	//WARN 只能用包装类，不能用原生类型,否则无法通过反射找到get、set方法，持久层会出错
 	private Integer id;
 	private String value;
-	private Float idf;
+	private Double idf;
 
 	public Integer getId() {
 		return id;
@@ -22,11 +22,11 @@ public class Dict {
 		this.value = value;
 	}
 
-	public Float getIdf() {
+	public Double getIdf() {
 		return idf;
 	}
 
-	public void setIdf(Float idf) {
+	public void setIdf(Double idf) {
 		this.idf = idf;
 	}
